+++ feature/motion/build.gradle.kts
@@ -56,9 +56,6 @@ dependencies {
     implementation(libs.androidx.compose.ui.tooling.preview)
     implementation(libs.androidx.navigation.compose)
     implementation(libs.androidx.media)
-    implementation(libs.vico.core)
-    implementation(libs.vico.compose)
-    implementation(libs.vico.compose.m3)
     implementation(libs.ui.graphics)
     implementation(libs.hilt.android)
     implementation(libs.androidx.hilt.navigation.compose)
